Major change in used js hierarchy
Now most of the types used are native java types
instead of duplicate proxy types, the only remaining
ones are JSUndefined and JSEntry. JSObject is now an
interface as of now we need access to value by nth
key (to be removed in future). There are 3 implementations
of JSObject IndexedHashMap, IterativeHashMap, ArrayMap
(each has appropriate javadoc).

* remove redundant classes of the previous hierarchy
* remove tests for the redundant classes
* remove unneeded methods from JSTypesUtil
* remove JSParser
* introduce JSObject and JSSerializable interfaces
* add implementations of JSObject
  (IndexedHashMap, IterativeHashMap, ArrayMap)
* add JSEntry (basic implementation of Map.Entry for JSObject)
* migrate usages in JSTPConnection
* migrate usages in JSTPMessage
* migrate CallHandler and StateHandler
* migrate TCPTransport
* migrate Tests (JSTPConnection and TCPTransport)
* migrate ManualHandler
* split JSNativeParser parse methods to public and
  internal to avoid unnecessary checks
* update idea project files