compiler: allow to auto generate simple proxies
Introduce Proxy Call and Event annotations that
allow to describe simple proxy interface using which
compiler will generate its implementation. As of now
it is possible to specify default interface in Proxy
annotation or directly specify interface for each method.
By default method name will be used as remote method name,
by that can be overridden as a first param of Call/Event
annotation if there is only one param or second if custom
interface is also specified. Handler must be the last parameter
and is optional.